Topic 6 Completed
Create a REST API to create, update and remove meetings.
Note that a meeting has attendees, room and a time slot.
Attendees and rooms exists in the system and must be retrieved by an endpoint to create the meeting.
You must provide the following features: create meetings, query existing ones, cancel and update meetings.